+++ karman_app/lib/app_shell.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/cupertino.dart';
-import 'package:karman_app/pages/habits_page.dart';
-import 'package:karman_app/pages/pomodoro_page.dart';
-import 'package:karman_app/pages/settings_page.dart';
-import 'package:karman_app/pages/tasks_page.dart';
-import 'package:karman_app/pages/zen_page.dart';
+import 'package:karman_app/pages/habit/habits_page.dart';
+import 'package:karman_app/pages/pomodoro/pomodoro_page.dart';
+import 'package:karman_app/pages/setting/settings_page.dart';
+import 'package:karman_app/pages/tasks/tasks_page.dart';
+import 'package:karman_app/pages/zen/zen_page.dart';
 
 class AppShell extends StatelessWidget {
   @override
